* fixed opengl compilation on Windows  * opengl version bump to 0.6.7

+++ src/modules/opengl/shader.c
@@ -68,7 +68,7 @@ API GLuint createOpenGLShaderFromString(const char *source, GLenum type)
 API GLuint createOpenGLShaderFromFile(char *filename, GLenum type)
 {
 	char *shaderSource;
-	unsigned long shaderLength;
+	gsize shaderLength;
 	if(!g_file_get_contents(filename, &shaderSource, &shaderLength, NULL)) {
 		LOG_ERROR("Failed to read shader source from %s", filename);
 		return false;
